examples/tldr/pyproject.toml
[project]
name = "tldr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain==0.3.27",
    "python-dotenv==1.0.1",
    "langchaingang[all]",
]

[tool.uv.sources]
langchaingang = { workspace = true }
